Take some of sys/vxworks/process/* from sys/unix instead.
