+++ library/std/src/sys/vxworks/process/mod.rs
@@ -1,7 +1,9 @@
-pub use self::process_common::{Command, ExitCode, ExitStatus, Stdio, StdioPipes};
-pub use self::process_inner::Process;
+pub use self::process_common::{Command, CommandArgs, ExitCode, Stdio, StdioPipes};
+pub use self::process_inner::{ExitStatus, Process};
 pub use crate::ffi::OsString as EnvKey;
+pub use crate::sys_common::process::CommandEnvs;
 
+#[path = "../../unix/process/process_common.rs"]
 mod process_common;
 #[path = "process_vxworks.rs"]
 mod process_inner;
